+++ NEWS.md
@@ -9,6 +9,8 @@
 * `system_check()` combines arguments with ` `, not `, `. (#753)
 
 * `lint()` gains a `cache` argument (@jimhester, #708).
+* `test()` gains an `...` argument so that additional arguments can be passed
+  to `testthat::test_dir` (@jimhester, #747)
 
 * export functions `RCMD()` and `system_check()` so they can be used by other 
   packages. (@jimhester, #699).